Break out reusable PersonSelect component from CampaignDetailsForm
