+++ src/components/forms/CampaignDetailsForm.tsx
@@ -1,15 +1,14 @@
 import { Form } from 'react-final-form';
+import { TextField } from 'mui-rff';
 import { useQuery } from 'react-query';
 import { useRouter } from 'next/router';
-import { Autocomplete, TextField } from 'mui-rff';
-import { Avatar, Box, Link, MenuItem, Typography } from '@material-ui/core';
+import { useState } from 'react';
+import { Link, MenuItem } from '@material-ui/core';
 import { FormattedMessage as Msg, useIntl } from 'react-intl';
-import { useEffect, useState } from 'react';
 
-import getPeopleSearchResults from 'fetching/getPeopleSearchResults';
 import getUserMemberships from 'fetching/getUserMemberships';
+import PersonSelect from './common/PersonSelect';
 import SubmitCancelButtons from './common/SubmitCancelButtons';
-import useDebounce from 'hooks/useDebounce';
 import { ZetkinCampaign, ZetkinPerson } from 'types/zetkin';
 
 interface CampaignDetailsFormProps {
@@ -26,41 +25,11 @@ const CampaignDetailsForm = ({ onSubmit, onCancel, campaign }: CampaignDetailsFo
     const activeMembership = membershipsQuery?.data?.find(m => m.organization.id.toString() == orgId);
     const userProfile = activeMembership?.profile;
 
-    const [searchFieldValue, setSearchFieldValue] = useState<string>('');
-
     const [selectedManager, setSelectedManager] = useState<Partial<ZetkinPerson> | null>(campaign?.manager? {
         first_name: campaign.manager.name.split(' ')[0],
         id: campaign.manager.id,
         last_name: campaign?.manager.name.split(' ')[1],
     } as Partial<ZetkinPerson> : null);
-
-    const { isLoading, refetch, data: results } = useQuery(
-        ['peopleSearchResults', searchFieldValue],
-        getPeopleSearchResults(searchFieldValue, orgId as string),
-        { enabled: false },
-    );
-
-    let searchLabel = searchFieldValue.length ?
-        intl.formatMessage({ id: 'misc.formDialog.campaign.keepSearch' }) :
-        intl.formatMessage({ id: 'misc.formDialog.campaign.search' });
-
-    if (isLoading) {
-        searchLabel = intl.formatMessage({ id: 'misc.formDialog.campaign.searching' });
-    }
-    else if (results?.length == 0) {
-        searchLabel = intl.formatMessage({ id: 'misc.formDialog.campaign.noResult' });
-    }
-
-    const debouncedQuery = useDebounce(async () => {
-        refetch();
-    }, 600);
-
-    // Watch for changes on the search field value and debounce search if changed
-    useEffect(() => {
-        if (searchFieldValue.length >= 3) {
-            debouncedQuery();
-        }
-    }, [searchFieldValue.length, debouncedQuery]);
 
     const initialValues = {
         info_text: campaign?.info_text,
@@ -87,11 +56,6 @@ const CampaignDetailsForm = ({ onSubmit, onCancel, campaign }: CampaignDetailsFo
             visibility,
         });
     };
-
-    let managerOptions = (results || []) as Partial<ZetkinPerson>[];
-    if (selectedManager && !managerOptions.some(o => o.id === selectedManager.id)) {
-        managerOptions = [selectedManager].concat(managerOptions);
-    }
 
     return (
         <Form
@@ -122,35 +86,15 @@ const CampaignDetailsForm = ({ onSubmit, onCancel, campaign }: CampaignDetailsFo
                         variant="outlined"
                     />
 
-                    <Autocomplete
-                        filterOptions={ (options) => options } // override filtering
-                        getOptionLabel={ person => person.first_name ? `${person.first_name} ${person.last_name}` : '' }
-                        getOptionSelected={ (option, value) => option?.id == value?.id }
-                        getOptionValue={ person => person.id || null }
-                        label={ intl.formatMessage({ id: 'misc.formDialog.campaign.manager.label' }) }
+                    <PersonSelect
                         name="manager_id"
-                        noOptionsText={ searchLabel }
-                        onChange={ (_, v) => {
-                            setSelectedManager(v as Partial<ZetkinPerson>);
+                        onChange={ (person) => {
+                            setSelectedManager(person);
                         } }
-                        onInputChange={ (_, v) => {
-                            setSearchFieldValue(v);
-                        } }
-                        options={ managerOptions }
-                        renderOption={ (person) => (
-                            <Box alignItems="center" display="flex">
-                                <Box m={ 1 }>
-                                    <Avatar
-                                        src={ `/api/orgs/${orgId}/people/${person.id}/avatar` }>
-                                    </Avatar>
-                                </Box>
-                                <Typography>
-                                    { `${ person.first_name } ${ person.last_name }` }
-                                </Typography>
-                            </Box>
-                        ) }
-                        value={ selectedManager }
+                        orgId={ orgId as string }
+                        selectedPerson={ selectedManager }
                     />
+
                     { (userProfile && userProfile.id != selectedManager?.id) && (
                         <Link
                             color="textPrimary"
